+++ front/src/components/pages/Home/AboutSummary.tsx
