+++ front/src/components/pages/Home/Banner.tsx
