+++ front/src/components/pages/Home/ProjectSummary.tsx
